Thread Names
Summary:
Add `SetThreadName` function and call it from a bunch of threads.

Note: You must name threads from within the thread with this interface. While most platforms support doing it from inside or outside of threads, Apple decided to be restrictive/bad, so I'm just making that a restriction for us so that it works on all platforms.

Test Plan: Run in debugger and look at thread names

Reviewers: mike.naquin

Reviewed By: mike.naquin

CC: cbrown

Differential Revision: https://live-phabricator.bittorrent.com/D673

[This commit was only partially migrated from livelib commit c6d44951af725bbe2e3f8795af1423c28f02eaa6.]