Minimal text field
Summary: No cursor, highlighting, multiline, not skin-able/customizable, yet. Currently should be the minimum we need for username/password fields - I'm going to move on to another task but if I have more time I'll start adding more features to this and make it more generalized.

Test Plan: add a TextFieldView to an existing view and use it (click on it to gain focus, and start typing/copy/paste/backspace)

Reviewers: cbrown

Reviewed By: cbrown

Subscribers: mike.naquin

Differential Revision: https://live-phabricator.bittorrent.com/D731