tvOS
Summary:
Mainly the only thing in our code that differs is the VideoToolbox stuff. Other changes were to keep the new clang compiler happy.

Also, I ended up forking SDL so that we could add some #ifdefs so that it compiles for tvOS. This was the majority of the work.

Test Plan: manual

Reviewers: cbrown

Subscribers: dapirian

Differential Revision: https://phabricator.watchonair.tv/D994

[This commit was only partially migrated from livelib commit 755efe804565b80c37023087161272fec09515a3.]