Xcode unit tests
Summary:
This will allow us to run our google tests inside of Xcode. The major advantage of this is being able to run our existing tests on actual devices, and within the simulators.

We can also use the built-in Xcode unit testing integration. This allows us to run a specific test with one click, and we don't have to mess with any command line options.

There was a bit of trickery to get this to work. In order to run on a device, we have to install a "host" app. This is just an empty app that links against our unit tests. The other bit is a little Google Test -> Xcode Unit Test file that converts each unit test into an Xcode unit test (see third-party/gtest-xcode/src/GoogleTest.mm). Once we have those, we just need an Xcode Unit Test bundle for each project (at some point I'd like a script to automate this part).

To keep the builds managable in Xcode, I added a libtool step to the build. We'll have two sets of combined libraries.

- bin/<platform>/<configuration>/libthird-party.a
- bin/<platform>/<configuration>/libcombined.a

The former is all of our 3rd-party libraries combined into one. The latter is all of our libaries combine into one.

Having these reduces the number of settings that we have to maintain in Xcode drastically.

I made some changes no needy so that it mostly will accept Xcode-friendly platform names. I also added a "Mac OS X" platform so that we can properly pass in the targeted macosx version. I'll submit this as a pull request in GitHub.

Test Plan: Unit Tests

Reviewers: cbrown

Subscribers: dapirian

Differential Revision: https://phabricator.watchonair.tv/D1039

[This commit was only partially migrated from livelib commit 4783817701ffae21534b328f0ba778028812e612.]