Update cppformat
Summary:
Update cppformat

So the only reason this (and the api-client-cpp one) are reviews and I didn't just do it is because I removed it from being a submodule and made them both just include the source directly. Their repo changed the directory hierarchy and it would no longer be a nice simple submodule setup.

But, I think Mike is moving this to needy (we should), so really this is temporary to get our logging working in the meantime.

Test Plan: manual

Reviewers: cbrown

Reviewed By: cbrown

Subscribers: #engineering

Differential Revision: https://phabricator.watchonair.tv/D1238

[This commit was only partially migrated from livelib commit 08febf1a95a7cb0fb462174afb0d7eaea9e22fd3.]