fix natural scroll direction
Summary: scrolling was written specifically for macbook trackpads before. the direction of the wheel event on a ball or wheel mouse in the direction that the finger moved. on the trackpad, it's inverted. this checks for that inversion flag from sdl and fixes the scroll view direction

Test Plan: scroll

Reviewers: vmrob

Reviewed By: vmrob

Subscribers: #engineering

Differential Revision: https://phabricator.watchonair.tv/D1267