Don't render views that are fully clipped
Summary: Added an early out. Could add another in `renderAndRenderSubviews` (not to be confused with `_renderAndRenderSubviews`) but that was more complicated. This isn't all we need to do, but it's the bulk of it, I think. Haven't tested on appletv, looks good on mac. Raises the lowest framerate moment from 26 to 45, but it's hard to tell much on mac since it almost always sits at 60.

Test Plan: manual

Reviewers: cbrown

Reviewed By: cbrown

Subscribers: #engineering

Differential Revision: https://phabricator.btlive.tv/D1576