Another empty clip bounds optimization
Summary:
Same thing, but for cached views

Finally got it running on appletv, with this and the other early out, it's still not nearly fast enough. When scrolling we drop to 3-4fps for a frame. Otherwise the fps is okay. It's not entirely unusable, but not shippable either.

Test Plan: use app

Reviewers: cbrown

Reviewed By: cbrown

Subscribers: vmrob

Differential Revision: https://phabricator.btlive.tv/D1580