Fix window assert on shutdown
Summary: We get a window close event when we want to close which triggers window.close(), which dispatches a window change with nullptr as the window. That's all fine, but afterwards when we destroy the window object we call close() again (on an already closed window) which dispatches a window change with nullptr again which triggers the assert.

Test Plan: close mac app normally, confirm no assert on shutdown anymore

Reviewers: cbrown

Reviewed By: cbrown

Subscribers: #engineering

Differential Revision: https://phabricator.btlive.tv/D1584