Fix build variants and cleanup configure
Summary:
Previously, the build variant selection wasn't being passed into the needs renderer due to lack of functionality in needy. Needy has been updated to provide this functionality and it is now utilized in configure.

Additionally, the build has been simplified with a simplified build-deps and configure step as well as an intermediate configure.json that is used to pass configuration from build-deps to configure.

Test Plan: travis

Reviewers: cbrown

Reviewed By: cbrown

Subscribers: #engineering

Differential Revision: https://phabricator.btlive.tv/D1857