Add error handling to utf8 decode
Summary: One of our strings was causing this thing to crash, I think because it was invalid utf8. We should just use a better library for this that detects and reports errors. I don't have a lot of faith in this one, but at least it won't crash anymore.

Test Plan: manual

Reviewers: vmrob

Reviewed By: vmrob

Subscribers: #engineering

Differential Revision: https://phabricator.btlive.tv/D1921