add okuiml text attributes and image element
Summary:
This adds several attributes to the <text> element, hex code color attribute parsing, and
the image element. To better facilitate the setting of image texture and placeholder independently,
the ImageView::setTextureFromURL method was split into setTexture and setTexturePlaceholder
methods. The previous setTexture's functionality was merged into the new one, so the same method is
used for resource and URL textures.

Test Plan: unit / app

Reviewers: vmrob

Reviewed By: vmrob

Subscribers: #engineering

Differential Revision: https://phabricator.btlive.tv/D1956